+++ transferservice/src/main/java/com/transferservice/dto/ConversionRateDto.java
@@ -1,5 +1,7 @@
 package com.transferservice.dto;
 
+import java.math.BigDecimal;
+
 public record ConversionRateDto(
-    Double rate
+    BigDecimal rate
 ) {}
